refactor: add improvements after review

+++ src/helpers/api.helper.ts
@@ -1,6 +1,7 @@
 import {expect, request} from '@playwright/test';
 
 import envConfig from '@_src/config/envConfig';
+import {ApiResponseStatuses} from '@_src/enums/apiResponseStatuses.enum';
 import {UserDataInterface} from '@_src/interfaces/userData.interface';
 
 export async function authenticateAndGetBearerToken(userData: UserDataInterface): Promise<string> {
@@ -11,24 +12,39 @@ export async function authenticateAndGetBearerToken(userData: UserDataInterface)
             password: userData.password,
         },
     });
-
     const responseJSON = await response.json();
-    expect(response.status()).toBe(200);
-    expect(responseJSON).toHaveProperty('access_token');
-
-    return responseJSON.access_token;
+    if (response.status() === 200) {
+        expect(response.statusText()).toBe(ApiResponseStatuses.OK);
+        expect(responseJSON).toHaveProperty('access_token');
+        return responseJSON.access_token;
+    } else {
+        expect(response.status()).toBe(401);
+        expect(response.statusText()).toBe(ApiResponseStatuses.Unauthorized);
+    }
 }
 
-export async function checkIfUserExistsByID(userID: string): Promise<void> {
-    const getUserEndpoint = envConfig.API.verifyUserExistsAPI.replace('{{id}}', userID);
+export async function checkIfUserExistsByID(userData: UserDataInterface): Promise<boolean> {
+    const getUserEndpoint = envConfig.API.verifyUserExistsAPI.replace('{{id}}', userData.id);
     const requestContext = await request.newContext();
-    const response = await requestContext.head(getUserEndpoint);
+    const response = await requestContext.get(getUserEndpoint);
 
-    expect(response.status()).toBe(200);
+    if (response.status() === 200) {
+        expect(response.statusText()).toBe(ApiResponseStatuses.OK);
+        return true;
+    } else if (response.status() === 404) {
+        expect(response.statusText()).toBe(ApiResponseStatuses.UserNotFound);
+        return false;
+    } else {
+        throw new Error('Invalid user ID is provided.');
+    }
 }
 
-export async function deleteUserWithAPI(userID: string, token: string): Promise<void> {
-    const deleteUserEndpoint = envConfig.API.deleteUserAPI.replace('{{id}}', userID);
+export async function deleteUserWithAPI(userData: UserDataInterface, token: string): Promise<void> {
+    if (!token) {
+        throw new Error('Access token is required but was not provided.');
+    }
+
+    const deleteUserEndpoint = envConfig.API.deleteUserAPI.replace('{{id}}', userData.id);
     const requestContext = await request.newContext();
     const response = await requestContext.delete(deleteUserEndpoint, {
         headers: {
@@ -37,5 +53,6 @@ export async function deleteUserWithAPI(userID: string, token: string): Promise<
     });
 
     expect(response.status()).toBe(200);
-    expect(response.statusText()).toBe('OK');
+    expect(response.statusText()).toBe(ApiResponseStatuses.OK);
+    console.info(`User ${userData.id} successfully deleted.`);
 }
